+++ lib/models/BMCommonCardModel.dart
@@ -9,5 +9,14 @@ class BMCommonCardModel {
   bool saveTag;
   bool? liked;
 
-  BMCommonCardModel({required this.image, required this.title, this.subtitle, this.comments, this.distance, this.likes, this.rating, required this.saveTag, this.liked});
+  BMCommonCardModel(
+      {required this.image,
+      required this.title,
+      this.subtitle,
+      this.comments,
+      this.distance,
+      this.likes,
+      this.rating,
+      required this.saveTag,
+      this.liked});
 }
